python/requirements.txt
absl-py==0.7.1
alabaster==0.7.12
appnope==0.1.0
astor==0.8.0
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
bidict==0.18.0
bleach==3.1.0
bokeh==1.3.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
docutils==0.14
entrypoints==0.3
fastdtw==0.3.2
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
icc-rt==2019.0
idna==2.8
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==0.18
intel-openmp==2019.0
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
jedi==0.14.1
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
llvmlite==0.29.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
notebook==6.0.0
numba==0.45.0
numpy==1.17.0rc2
packaging==19.0
pandas==0.25.0
pandocfilters==1.4.2
parso==0.5.1
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
pluggy==0.12.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.0
ptyprocess==0.6.0
py==1.8.0
Pygments==2.4.2
PyOpenGL==3.1.3b2
pyparsing==2.4.1
PyQt5==5.12.1
PyQt5-sip==4.19.18
PyQt5-stubs==5.12.1.0
pyqtgraph==0.10.0
pyrsistent==0.15.3
pytest==5.0.1
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1.1
pyzmq==18.0.2
qtconsole==4.5.2
requests==2.22.0
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
snowballstemmer==1.9.0
Sphinx==2.1.2
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
statsmodels==0.10.1
tb-nightly==1.14.0a20190603
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tf-estimator-nightly==1.14.0.dev2019060501
tornado==6.0.3
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.5.0
wrapt==1.11.2
xarray==0.12.3
zipp==0.5.2
